requirements.txt
streamlit
python-docx
pandas
openpyxl
requests
beautifulsoup4
python-dotenv
pdfplumber
pymupdf
matplotlib
plotly
duckduckgo-search
google-genai
Pillow
pytest
pytest-cov
# Add other dependencies as needed
